refactor: shape class refactor
