+++ OSiSP_labs/Lab01/Lab01.vcxproj
@@ -127,7 +127,9 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
+    <ClCompile Include="drawer.cpp" />
     <ClCompile Include="main.cpp" />
+    <ClCompile Include="shape.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="resource.h" />
